Fix ReLU using Double

+++ core-entities/src/commonMain/kotlin/org/pointyware/ember/entities/activations/ReLU.kt
@@ -4,11 +4,11 @@ package org.pointyware.ember.entities.activations
  * Rectified Linear Unit (ReLU) activation function.
  */
 object ReLU: ScalarActivationFunction {
-    override fun scalarActivation(input: Double): Double {
-        return if (input > 0) input else 0.0
+    override fun scalarActivation(input: Float): Float {
+        return if (input > 0) input else 0.0f
     }
 
-    override fun scalarDerivative(input: Double): Double {
-        return if (input > 0) 1.0 else 0.0
+    override fun scalarDerivative(input: Float): Float {
+        return if (input > 0) 1.0f else 0.0f
     }
 }
